requirements.txt
openai
langchain
streamlit
requests
python-dotenv
chromadb
